gets user names

+++ main.go
@@ -2,24 +2,32 @@ package main
 
 import (
 	"fmt"
-	"hypixel-auction-v4/database"
+	"hypixel-auction-v4/HypixelRequests/auctions"
+	database "hypixel-auction-v4/MongoDatabase"
 	"os"
 	"time"
 )
 
 func main() {
 
-	start := time.Now()
-	if os.Getenv("UseDB") == "true" {
-		database.RemoveAll()
-		database.Test()
+	time.Sleep(10 * time.Second)
 
-		for {
-			time.Sleep(10 * time.Second)
-			database.UpdateData()
+	start := time.Now()
+	if os.Getenv("test") != "" {
+		if os.Getenv("UseDB") == "true" {
+			database.RemoveAll()
+			database.Test()
+
+			for {
+				time.Sleep(10 * time.Second)
+				database.UpdateData()
+			}
+			// database.Test()
 		}
-		// database.Test()
 	}
+
+	auctions.AllPagesAuctions(time.Now())
+
 	fmt.Println(time.Since(start))
 	// fmt.Println(auctions.URL)
 	// auctions.AllPagesAuctions()
